+++ MyChat/Aplication/Base.lproj/LaunchScreen.storyboard
