+++ MyChat/Aplication/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
