+++ MyChat/Aplication/Resources/Assets.xcassets/Contents.json
